+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * Implement ALSA underrun (#371)
 * Implement Steam callback mechanism
 * Implement Steam Networking
+* Implement more of Steam remote storage
 
 ### Changed